fix issue about the status when create the user

+++ src/main/java/com/task/constants/MainConstants.java
@@ -6,7 +6,7 @@ public class MainConstants {
     
 	public static final long ACCOUNT_ACTIVE = 1;
 	public static final int ACCOUNT_SUSPENDED = 2;
-	public static final int ACCOUNT_LOCKED = -1;
+	public static final long ACCOUNT_LOCKED = -1;
     public static final long JWT_SECRET_KEY = 1;
 	
 }
